# packages/api-contracts/openapi/platform-api.yaml
openapi: 3.1.0
info:
  title: Actenora Platform API
  version: 0.2.0
  description: Platform + web-portal contract surface for the modular monolith (FAZ 24).
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/approvals/{approvalId}/decide:
    post:
      operationId: decideApproval
      summary: Approve, reject, or request changes on an approval request
      parameters:
        - name: approvalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionRequest"
      responses:
        "200":
          description: Updated approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRecord"
  /api/v1/meeting-notes/{noteId}/submit-for-approval:
    post:
      operationId: submitMeetingNoteForApproval
      summary: Submit a draft meeting note version for human approval
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitForApprovalRequest"
      responses:
        "200":
          description: Approval opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRecord"
  /api/v1/approvals/{approvalId}:
    get:
      operationId: getApproval
      summary: Fetch approval request status
      parameters:
        - name: approvalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Approval snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRecord"
...
components:
...
  schemas:
...
    ApprovalRecord:
      type: object
      additionalProperties: false
      required: [id, artifactType, artifactId, status]
      properties:
        id:
          type: string
          format: uuid
        artifactType:
          type: string
        artifactId:
          type: string
          format: uuid
        noteId:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [PENDING, GRANTED, DENIED, EXPIRED, CHANGES_REQUESTED]
        version:
          type: integer
          format: int64
        updatedAt:
          type: string
          format: date-time
          nullable: true
        decidedBy:
          type: string
          nullable: true
        decidedAt:
          type: string
          format: date-time
          nullable: true
        comment:
          type: string
          nullable: true
    ApprovalDecisionRequest:
      type: object
      additionalProperties: false
      required: [decision]
      properties:
        decision:
          type: string
          enum: [APPROVE, REJECT, REQUEST_CHANGES]
        comment:
          type: string
        expectedNoteVersion:
          type: integer
          format: int64
        expectedApprovalVersion:
          type: integer
          format: int64
    SubmitForApprovalRequest:
      type: object
      additionalProperties: false
      properties:
        approverId:
          type: string
        expiresAt:
          type: string
          format: date-time
        expectedVersion:
          type: integer
          format: int64
...